+++ Pandora/Paradigm/Structure/Ability.hs
@@ -2,5 +2,6 @@ module Pandora.Paradigm.Structure.Ability (module Exports) where
 
 import Pandora.Paradigm.Structure.Ability.Zipper as Exports
 import Pandora.Paradigm.Structure.Ability.Substructure as Exports
+import Pandora.Paradigm.Structure.Ability.Rotatable as Exports
 import Pandora.Paradigm.Structure.Ability.Focusable as Exports
 import Pandora.Paradigm.Structure.Ability.Nonempty as Exports