+++ CHANGELOG.md
@@ -242,3 +242,4 @@
 * Define left and right `zig`, `zig-zig` and `zig-zag` `Splay` rotations for `Binary` tree
 * Define experimental methods in `Extendable` class - `$=>>` and `<<=$`
 * Define `substructure` method which doesn't involve `Tagged`
+* Define `Rotatable` typeclass to rotate trees